+++ IssueTracker/IssueTracker/Extension/UIStackView+Extension.swift
@@ -1,0 +1,10 @@
+import UIKit
+
+extension UIStackView {
+    func clearSubviews() {
+        self.arrangedSubviews.forEach { view in
+            self.removeArrangedSubview(view)
+            view.removeFromSuperview()
+        }
+    }
+}
