+++ IssueTracker/IssueTracker/Issue/View/IssueListCell.swift
@@ -1,13 +1,5 @@
 import UIKit
 
-extension UIStackView {
-    func clearSubviews() {
-        self.arrangedSubviews.forEach { view in
-            self.removeArrangedSubview(view)
-            view.removeFromSuperview()
-        }
-    }
-}
 
 final class IssueListCell: UICollectionViewCell {
